requirements.txt
python-telegram-bot==20.3
requests==2.31.0
numpy==1.26.0
scikit-learn==1.3.2
openai==1.31.0
pdfplumber==0.9.0
sentence-transformers
transformers